Proseguimento generale
Aggiunti nuovi layout e rispettivi controller (ancora da implementare), modificato del codice, creato nuove classi ausiliarie, spostate e/o rinominate alcune classi.

+++ source-code/Client/frontendandroid/app/src/main/java/com/iasdietideals24/dietideals24/controller/ControllerCreazioneProfiloFase1.kt
@@ -167,7 +167,7 @@ class ControllerCreazioneProfiloFase1 : Controller(R.layout.creazioneprofilofase
             .setTitleText(R.string.creazioneProfiloFase1_titoloPopupData)
             .setCalendarConstraints(calendarConstraints.build())
             .setSelection(
-                if (viewModel.dataNascita.value?.time == (0.toLong()))
+                if (viewModel.dataNascita.value?.time == 0L)
                     MaterialDatePicker.todayInUtcMilliseconds()
                 else
                     viewModel.dataNascita.value?.time
